third_party/blink/web_tests/platform/generic/fast/forms/date/date-click-on-label-expected.txt

Check if clicking a label focuses on a date input in it.
PASS document.activeElement.id is "body"
Clicking a label.
PASS document.activeElement.id is "date1"
PASS successfullyParsed is true

TEST COMPLETE

Label: 
